requirements.txt
asgiref==3.7.2
cfgv==3.4.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.13.1
identify==2.5.33
inflection==0.5.1
MarkupPy==1.14
nodeenv==1.8.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
platformdirs==4.2.0
pre-commit==3.6.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
virtualenv==20.25.0
xlrd==2.0.1
xlwt==1.3.0
